// src/guitar/models/interval.enum.ts
export const enum Interval {
  PERFECT_UNISON = 0,
  MINOR_SECOND = 1,
  SEMITONE = 1,
  MAJOR_SECOND = 2,
  TONE = 2,
  MINOR_THIRD = 3,
  MAJOR_THIRD = 4,
  PERFECT_FORTH = 5,
  DIMINISHED_FIFTH = 6,
  AUGMENTED_FOURTH = 6,
  TRITONE = 6,
  PERFECT_FIFTH = 7,
  MINOR_SIXTH = 8,
  MAJOR_SIXTH = 9,
  MINOR_SEVENTH = 10,
  MAJOR_SEVENTH = 11,
  PERCECT_OCTAVE = 12
}